reorganize to support CLI and lib/API usage
